requirements.txt
gradio
plotly
pandas
python-dotenv
requests
transformers
torch
datasets
scikit-learn
psycopg2-binary
yfinance